[tools] Add drake_lcm_cc_library opt-in for new lcm_gen
Manual testing of flipping the default for this flag pointed out some
embarrassing problems with the new tool, which are also fixed here:

- Messages with sub-structs must include the headers for those structs.

- The getTypeName function is actually required downstream.

- Generated filenames must be `*.hpp` not `*.h`. Ideally we would have
  started a new filename convention with the new tool, but so much old
  code expects the messages to keep their old names that renaming is
  intractable in practice.
