+++ lcmtypes/BUILD.bazel
@@ -3,6 +3,7 @@ package(default_visibility = ["//visibility:public"])
 load("//tools/install:install.bzl", "install")
 load(
     "//tools/skylark:drake_cc.bzl",
+    "drake_cc_googletest",
     "drake_cc_library",
     "drake_transitive_installed_hdrs_filegroup",
 )
@@ -21,6 +22,7 @@ load(
     "drake_lcm_java_library",
     "drake_lcm_py_library",
 )
+load("//tools/workspace:generate_file.bzl", "generate_file")
 load(":defs.bzl", "ALL_LCM_SRCS")
 load("//tools/lint:lint.bzl", "add_lint_tests")
 
@@ -380,6 +382,34 @@ install(
 )
 
 # === test/ ===
+
+# TODO(jwnimmer-tri) We use this to test the `use_new_lcm_gen = True` flag.
+# Once we turn on that flag by default, we can probably ditch this test.
+generate_file(
+    name = "test_message.lcm",
+    content = """
+package drake;
+struct test_message { double value; }
+""",
+    visibility = ["//visibility:private"],
+)
+
+drake_lcm_cc_library(
+    name = "test_message",
+    lcm_package = "drake",
+    lcm_srcs = ["test_message.lcm"],
+    use_new_lcm_gen = True,
+    visibility = ["//visibility:private"],
+)
+
+drake_cc_googletest(
+    name = "test_message_test",
+    deps = [
+        ":test_message",
+        "//lcm:lcm_messages",
+    ],
+)
+
 drake_py_test(
     name = "nested_types_test",
     deps = [
